requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
arrow==0.15.2
asyncpg==0.30.0
attrs==25.4.0
cachetools==6.2.2
certifi==2025.11.12
cfgv==3.4.0
charset-normalizer==3.4.4
Click==7.0
coloredlogs==15.0.1
cursor==1.3.5
daily-python==0.22.0
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.121.2
filelock==3.20.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==1.7.11
google-auth==2.43.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-speech==2.34.0
google-cloud-texttospeech==2.33.0
google-genai==1.51.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
gspread==3.1.0
h11==0.16.0
halo==0.0.28
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
identify==2.6.15
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
llvmlite==0.44.0
log-symbols==0.0.14
loguru==0.7.3
Markdown==3.10
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
nltk==3.9.2
nodeenv==1.9.1
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.8.1
packaging==25.0
pillow==11.3.0
pipecat-ai==0.0.91
platformdirs==4.5.0
pre_commit==4.4.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydantic[email]
PyJWT
pyloudnorm==0.1.1
pyparsing==3.2.5
python-dateutil==2.8.0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
resampy==0.4.3
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
slackclient==2.2.1
soxr==0.5.0.post1
spinners==0.0.24
SQLAlchemy==2.0.44
starlette==0.49.3
sympy==1.14.0
tenacity==9.1.2
termcolor==3.2.0
threadpoolctl==3.6.0
tokenizers==0.22.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
Unidecode==1.4.0
uritemplate==3.0.1
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.35.4
websockets==15.0.1
win32_setctime==1.2.0
yarl==1.22.0

